improve testing
