+++ src/components/shared/Dialog.svelte
@@ -22,7 +22,7 @@
     >
       {text}
     </Dialog.Trigger>
-    <Dialog.Content class="max-h-11/12 w-[95%] max-w-250">
+    <Dialog.Content class="max-h-11/12 w-[95%] max-w-250 bg-card">
       {@render children()}
     </Dialog.Content>
   </form>
